requirements.txt
google-generativeai>=0.3.0
langchain>=0.1.0
langgraph>=0.0.20
langsmith>=0.0.75
sqlalchemy>=2.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=7.0.0
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4
numpy==1.26.2
